refactor helm check to base
