+++ generators/kubernetes-helm/index.js
@@ -23,7 +23,7 @@ const prompts = require('../kubernetes/prompts');
 const writeFiles = require('./files').writeFiles;
 const BaseDockerGenerator = require('../generator-base-docker');
 const { checkImages, generateJwtSecret, configureImageNames, setAppsFolderPaths } = require('../docker-base');
-const { checkKubernetes, loadConfig, saveConfig, setupKubernetesConstants, setupHelmConstants } = require('../kubernetes-base');
+const { checkKubernetes, checkHelm, loadConfig, saveConfig, setupKubernetesConstants, setupHelmConstants } = require('../kubernetes-base');
 const statistics = require('../statistics');
 
 module.exports = class extends BaseDockerGenerator {
@@ -35,24 +35,7 @@ module.exports = class extends BaseDockerGenerator {
             },
             ...super.initializing,
             checkKubernetes,
-            checkHelm() {
-                if (this.skipChecks) return;
-                const done = this.async();
-
-                shelljs.exec(
-                    'helm version --client | grep -E "v2\\.1[2-9]{1,2}\\.[0-9]{1,3}"',
-                    { silent: true },
-                    (code, stdout, stderr) => {
-                        if (stderr || code !== 0) {
-                            this.log(
-                                `${chalk.yellow.bold('WARNING!')} helm 2.12.x or later is not installed on your computer.\n` +
-                                    'Make sure you have helm installed. Read https://github.com/helm/helm/\n'
-                            );
-                        }
-                        done();
-                    }
-                );
-            },
+            checkHelm,
             loadConfig,
             setupKubernetesConstants,
             setupHelmConstants
